+++ docs/source/distributions.rst
@@ -274,13 +274,6 @@ Unit
     :undoc-members:
     :show-inheritance:
 
-VonMises
---------
-.. autoclass:: pyro.distributions.VonMises
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
 VonMises3D
 ----------
 .. autoclass:: pyro.distributions.VonMises3D
