+++ pyro/contrib/randomvariable/__init__.py
@@ -1,3 +1,6 @@
+# Copyright Contributors to the Pyro project.
+# SPDX-License-Identifier: Apache-2.0
+
 from pyro.contrib.randomvariable.random_variable import RandomVariable
 
 __all__ = [
