+++ pyro/distributions/torch.py
@@ -117,13 +117,6 @@ class LogNormal(torch.distributions.LogNormal, TorchDistributionMixin):
 class MultivariateNormal(torch.distributions.MultivariateNormal, TorchDistributionMixin):
     support = IndependentConstraint(constraints.real, 1)  # TODO move upstream
 
-    # TODO: remove this in the PyTorch release > 1.4.0
-    @lazy_property
-    def precision_matrix(self):
-        identity = torch.eye(self.loc.size(-1), device=self.loc.device, dtype=self.loc.dtype)
-        return torch.cholesky_solve(identity, self._unbroadcasted_scale_tril).expand(
-            self._batch_shape + self._event_shape + self._event_shape)
-
 
 class Normal(torch.distributions.Normal, TorchDistributionMixin):
     pass
